src/requirements.txt
--trusted-host files.pythonhosted.org --trusted-host pypi.org --trusted-host pypi.python.org
Pillow~=10.3
topshelfsoftware-util @ git+https://github.com/topshelfsoftware/python-utility.git@v1.0.0